chore: try to fix the target origin mismatch error
After adding the webform submission limit feature, we got this error:
  Failed to execute 'postMessage' on 'DOMWindow':
  The target origin provided ({FE-URL}) does not
  match the recipient window's origin ('{BE-URL}').

From the docs: https://iframe-resizer.com/troubleshooting/
  By default the origin of incoming messages is checked against the src
  attribute of the iframe.

Probably it does that only once. And then it does not change the origin
if the src has changed.

And in the recent change we started to use useState for the src with
empty string as the default.

So this commit should hopefully fix the issue by replacing the empty
string with some initial URL.